# Conflicts:
#	Package.resolved
#	Package.swift

+++ Sources/SwiftSDKGenerator/Generator/SwiftSDKGenerator+Copy.swift
@@ -5,8 +5,8 @@
 // Copyright (c) 2022-2023 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
-// See http://swift.org/LICENSE.txt for license information
-// See http://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+// See https://swift.org/LICENSE.txt for license information
+// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
 //
 //===----------------------------------------------------------------------===//
 
@@ -14,19 +14,15 @@ import SystemPackage
 
 extension SwiftSDKGenerator {
   func copyTargetSwiftFromDocker() async throws {
-    logGenerationStep("Building a Docker image for the target environment...")
-    // FIXME: launch Swift base image directly instead of building a new empty image
-    let imageName = try await buildDockerImage(baseImage: self.versionsConfiguration.swiftBaseDockerImage)
-
     logGenerationStep("Launching a Docker container to copy Swift SDK for the target triple from it...")
-    let containerID = try await launchDockerContainer(imageName: imageName)
+    let containerID = try await launchDockerContainer(imageName: self.versionsConfiguration.swiftBaseDockerImage)
     do {
       let pathsConfiguration = self.pathsConfiguration
 
       try await inTemporaryDirectory { generator, _ in
         let sdkUsrPath = pathsConfiguration.sdkDirPath.appending("usr")
         let sdkUsrLibPath = sdkUsrPath.appending("lib")
-        try generator.createDirectoryIfNeeded(at: sdkUsrPath)
+        try await generator.createDirectoryIfNeeded(at: sdkUsrPath)
         try await generator.copyFromDockerContainer(
           id: containerID,
           from: "/usr/include",
@@ -50,36 +46,46 @@ extension SwiftSDKGenerator {
             '
             """#
           )
-
-          let sdkUsrLib64Path = sdkUsrPath.appending("lib64")
-          try await generator.copyFromDockerContainer(
-            id: containerID,
-            from: FilePath("/usr/lib64"),
-            to: sdkUsrLib64Path
-          )
-
-          try createSymlink(at: pathsConfiguration.sdkDirPath.appending("lib64"), pointingTo: "./usr/lib64")
-
-          // `libc.so` is a linker script with absolute paths on RHEL, replace with a relative symlink
-          let libcSO = sdkUsrLib64Path.appending("libc.so")
-          try removeFile(at: libcSO)
-          try createSymlink(at: libcSO, pointingTo: "libc.so.6")
         }
 
-        try generator.createDirectoryIfNeeded(at: sdkUsrLibPath)
-        for subpath in ["clang", "gcc", "swift", "swift_static"] {
+        let sdkUsrLib64Path = sdkUsrPath.appending("lib64")
+        try await generator.copyFromDockerContainer(
+          id: containerID,
+          from: FilePath("/usr/lib64"),
+          to: sdkUsrLib64Path
+        )
+        try await createSymlink(at: pathsConfiguration.sdkDirPath.appending("lib64"), pointingTo: "./usr/lib64")
+
+        if case .rhel = self.versionsConfiguration.linuxDistribution {
+          // `libc.so` is a linker script with absolute paths on RHEL, replace with a relative symlink
+          let libcSO = sdkUsrLib64Path.appending("libc.so")
+          try await removeFile(at: libcSO)
+          try await createSymlink(at: libcSO, pointingTo: "libc.so.6")
+        }
+
+        try await generator.createDirectoryIfNeeded(at: sdkUsrLibPath)
+        var subpaths =  ["clang", "gcc", "swift", "swift_static"]
+
+        // Ubuntu's multiarch directory scheme puts some libraries in
+        // architecture-specific directories:
+        //   https://wiki.ubuntu.com/MultiarchSpec
+        if case .ubuntu = self.versionsConfiguration.linuxDistribution {
+          subpaths += ["\(targetTriple.cpu)-linux-gnu"]
+        }
+
+        for subpath in subpaths {
           try await generator.copyFromDockerContainer(
             id: containerID,
             from: FilePath("/usr/lib").appending(subpath),
             to: sdkUsrLibPath.appending(subpath)
           )
         }
+        try await generator.createSymlink(at: pathsConfiguration.sdkDirPath.appending("lib"), pointingTo: "usr/lib")
 
         // Python artifacts are redundant.
-        try generator.removeRecursively(at: sdkUsrLibPath.appending("python3.10"))
+        try await generator.removeRecursively(at: sdkUsrLibPath.appending("python3.10"))
 
-        try generator.createSymlink(at: pathsConfiguration.sdkDirPath.appending("lib"), pointingTo: "usr/lib")
-        try generator.removeRecursively(at: sdkUsrLibPath.appending("ssl"))
+        try await generator.removeRecursively(at: sdkUsrLibPath.appending("ssl"))
         try await generator.copyTargetSwift(from: sdkUsrLibPath)
         try await generator.stopDockerContainer(id: containerID)
       }
